feat(organization): Organize codebase
