+++ src/rest/mod.rs
@@ -1,0 +1,5 @@
+mod client;
+pub mod types;
+
+pub use client::KalshiRestClient;
+pub use types::*;
